+++ src/model/Settings/PortConfigurationSettings/PortConfigurationSettings.ts
@@ -12,8 +12,24 @@ export enum PortState {
 }
 
 export const DEFAULT_BAUD_RATES = [
-  110, 300, 1200, 2400, 4800, 9600, 14400, 19200, 38400, 57600, 115200,
+  110, 300, 1200, 2400, 4800, 9600, 14400, 19200, 38400, 57600, 115200, 230400, 460800, 921600
 ];
+
+export const NUM_DATA_BITS_OPTIONS = [5, 6, 7, 8];
+
+/**
+ * The string values must match that of the ParityType type provided
+ * by the Web Serial API.
+ */
+export enum Parity {
+  NONE = 'none',
+  EVEN = 'even',
+  ODD = 'odd',
+};
+
+export type StopBits = 1 | 1.5 | 2;
+
+export const STOP_BIT_OPTIONS: StopBits[] = [1, 2];
 
 const CONFIG_KEY = ['settings', 'port-configuration-settings'];
 
@@ -23,9 +39,15 @@ class Config {
    * This will cause the app to ignore whatever is in local storage and use the defaults,
    * updating to this new version.
    */
-  version = 1;
+  version = 2;
 
   baudRate = 115200;
+
+  numDataBits = 8;
+
+  parity = Parity.NONE;
+
+  stopBits: StopBits = 1;
 
   connectToSerialPortAsSoonAsItIsSelected = true;
 
@@ -46,7 +68,12 @@ export default class PortConfiguration {
 
   baudRateInputValue = this.config.baudRate.toString();
 
-  baudRateValidation = z.coerce.number().int().min(1).max(115200);
+  /**
+   * Set min. baud rate to 1 and max. baud rate to 2,000,000. Most systems won't actually
+   * support these ranges but let's not limit the user (they don't normally error if an unsupported
+   * number is provided, it just doesn't work correctly).
+   */
+  baudRateValidation = z.coerce.number().int().min(1).max(2000000);
   baudRateErrorMsg = '';
 
   constructor(appStorage: AppStorage) {
@@ -73,6 +100,24 @@ export default class PortConfiguration {
       // error message
       this.baudRateErrorMsg = parsed.error.errors[0].message;
     }
+  }
+
+  setNumDataBits = (numDataBits: number) => {
+    if (typeof numDataBits !== 'number') {
+      throw new Error("numDataBits must be a number");
+    }
+    this.config.numDataBits = numDataBits;
+    this._saveConfig();
+  }
+
+  setParity = (parity: Parity) => {
+    this.config.parity = parity;
+    this._saveConfig();
+  }
+
+  setStopBits = (stopBits: StopBits) => {
+    this.config.stopBits = stopBits;
+    this._saveConfig();
   }
 
   setConnectToSerialPortAsSoonAsItIsSelected = (value: boolean) => {
@@ -111,12 +156,31 @@ export default class PortConfiguration {
     // At this point we are confident that the deserialized config matches what
     // this classes config object wants, so we can go ahead and update.
     updateConfigFromSerializable(deserializedConfig, this.config);
+
+    this.setBaudRateInputValue(this.config.baudRate.toString());
   };
 
   _saveConfig = () => {
     const serializableConfig = createSerializableObjectFromConfig(this.config);
     this.appStorage.saveConfig(CONFIG_KEY, serializableConfig);
   };
+
+  /**
+   * Computed value which represents the serial port config in short hand,
+   * e.g. "115200 8n1"
+   *
+   * @returns The short hand serial port config for displaying.
+   */
+  get shortSerialConfigName() {
+    let output = '';
+    output += this.config.baudRate.toString();
+    output += ' ';
+    // output += this.selectedNumDataBits.toString();
+    // output += this.selectedParity[0];
+    // output += this.selectedStopBits.toString();
+    output += '8n1'; // TODO: Fix this
+    return output;
+  }
 }