+++ src/model/Settings/Settings.tsx
@@ -12,7 +12,7 @@ import GeneralSettings from './GeneralSettings/GeneralSettings';
 import AppStorage from '../Storage/AppStorage';
 import FakePortsController from 'src/model/FakePorts/FakePortsController';
 
-export type StopBits = 1 | 1.5 | 2;
+
 
 export enum SettingsCategories {
   PORT_CONFIGURATION,
@@ -40,27 +40,6 @@ export class Settings {
 
   generalSettings: GeneralSettings;
 
-  selectedPortPath = '';
-
-  // Commonly-available baud rates as mentioned at https://serialport.io/docs/api-stream/
-  baudRates = [
-    110, 300, 1200, 2400, 4800, 9600, 14400, 19200, 38400, 57600, 115200,
-  ];
-
-  selectedBaudRate = 115200;
-
-  numDataBitsOptions = [5, 6, 7, 8];
-
-  selectedNumDataBits = 8;
-
-  parityOptions = ['none', 'even', 'odd'];
-
-  selectedParity = 'none';
-
-  stopBitOptions: StopBits[] = [1, 2];
-
-  selectedStopBits: StopBits = 1;
-
   /**
    * Constructor for the Settings class.
    *
@@ -79,45 +58,13 @@ export class Settings {
     makeAutoObservable(this); // Make sure this is at the end of the constructor
   }
 
-  /**
-   * Computed value which represents the serial port config in short hand,
-   * e.g. "115200 8n1"
-   *
-   * @returns The short hand serial port config for displaying.
-   */
-  get shortSerialConfigName() {
-    let output = '';
-    output += this.selectedBaudRate.toString();
-    output += ' ';
-    output += this.selectedNumDataBits.toString();
-    output += this.selectedParity[0];
-    output += this.selectedStopBits.toString();
-    return output;
-  }
-
   setActiveSettingsCategory(settingsCategory: SettingsCategories) {
     this.activeSettingsCategory = settingsCategory;
   }
 
-  setSelectedPortPath(selectedPortPath: string) {
-    this.selectedPortPath = selectedPortPath;
-  }
-
-  setSelectedBaudRate(baudRate: number) {
-    this.selectedBaudRate = baudRate;
-  }
-
-  setSelectedNumDataBits(numDataBits: number) {
-    this.selectedNumDataBits = numDataBits;
-  }
-
-  setSelectedParity(parity: string) {
-    this.selectedParity = parity;
-  }
-
-  setSelectedStopBits(stopBits: StopBits) {
-    this.selectedStopBits = stopBits;
-  }
+  // setSelectedPortPath(selectedPortPath: string) {
+  //   this.selectedPortPath = selectedPortPath;
+  // }
 
   onKeyDown = (event: React.KeyboardEvent<HTMLDivElement>) => {
     // console.log(event);